+++ backend/src/main/kotlin/org/loculus/backend/controller/SubmissionController.kt
@@ -34,7 +34,6 @@ import org.loculus.backend.api.SubmissionIdFilesMap
 import org.loculus.backend.api.SubmissionIdMapping
 import org.loculus.backend.api.SubmittedProcessedData
 import org.loculus.backend.api.UnprocessedData
-import org.loculus.backend.api.getAllFileIds
 import org.loculus.backend.auth.AuthenticatedUser
 import org.loculus.backend.auth.HiddenParam
 import org.loculus.backend.config.BackendConfig
@@ -117,7 +116,9 @@ open class SubmissionController(
                 innerDataUseTermsType = dataUseTermsType
             }
         }
-        val fileMappingParsed = parseAndValidateFileMapping(fileMapping, groupId)
+        val fileMappingParsed = fileMapping?.let {
+            objectMapper.readValue(it, object : TypeReference<SubmissionIdFilesMap>() {})
+        }
         val params = SubmissionParams.OriginalSubmissionParams(
             organism,
             authenticatedUser,
@@ -144,7 +145,9 @@ open class SubmissionController(
         ) @RequestParam sequenceFile: MultipartFile?,
         @RequestPart fileMapping: String?,
     ): List<SubmissionIdMapping> {
-        val fileMappingParsed = parseAndValidateFileMapping(fileMapping, -1) // TODO what is the group ID here?
+        val fileMappingParsed = fileMapping?.let {
+            objectMapper.readValue(it, object : TypeReference<SubmissionIdFilesMap>() {})
+        }
         val params = SubmissionParams.RevisionSubmissionParams(
             organism,
             authenticatedUser,
@@ -153,29 +156,6 @@ open class SubmissionController(
             fileMappingParsed,
         )
         return submitModel.processSubmissions(UUID.randomUUID().toString(), params)
-    }
-
-    private fun parseAndValidateFileMapping(maybeFileMapping: String?, groupId: Int): SubmissionIdFilesMap? {
-        val fileMappingParsed = maybeFileMapping?.let {
-            objectMapper.readValue(it, object : TypeReference<SubmissionIdFilesMap>() {})
-        }
-        fileMappingParsed?.let {
-            val usedFileIds = getAllFileIds(it)
-            // Check if all given file IDs exist in the DB
-            val notExistingIds = filesDatabaseService.notExistingIds(usedFileIds)
-            if (notExistingIds.isNotEmpty()) {
-                throw BadRequestException("The File IDs $notExistingIds do not exist.")
-            }
-            // Check that all files belong to the given group
-            filesDatabaseService.getGroupIds(usedFileIds).forEach {
-                if (it.value != groupId) {
-                    throw BadRequestException(
-                        "The File ${it.key} belongs to group ${it.value} but should be long to group $groupId.",
-                    )
-                }
-            }
-        }
-        return fileMappingParsed
     }
 
     @Operation(description = EXTRACT_UNPROCESSED_DATA_DESCRIPTION)
